+++ NEWS.md
@@ -1,10 +1,15 @@
-# Items for next tagged release
+# cmdstanr 0.3.0.9000
 
 ### Bug fixes
 
 ### New features
 
-* `read_cmdstan_csv` now also returns time for MCMC sampling CSV files.
+* New function `as_cmdstan_fit()` that creates CmdStanMCMC/MLE/VB objects
+directly from CmdStan CSV files. (#412)
+
+* `read_cmdstan_csv` now also returns chain run times for MCMC sampling CSV
+files. (#414)
+
 
 # cmdstanr 0.3.0